app/pyproject.toml
[tool.poetry]
name = "rag-ib"
version = "0.1.0"
description = ""
authors = [" Nasibullin Alan Maratovich <alan.nasibyllin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
greenlet = "3.1.1"
pyee = "12.0.0"
typing-extensions = "4.12.2"
bm25s = "^0.2.5"
polars = "^1.17.1"
fastapi = "^0.115.6"
pydentic = "^0.0.1.dev3"
uvicorn = "^0.32.1"
pystemmer = "^2.2.0.3"
openai = "0.28"
python-dotenv = "^1.0.1"
feedparser = "^6.0.11"
httpx = "^0.28.1"
torch = {version = "^2.5.1+cpu", source = "torch_cpu"}
torchvision = {version = "^0.20.1+cpu", source = "torch_cpu"}
transformers = "^4.47.0"
numpy = "^2.2.0"


[[tool.poetry.source]]
name = "torch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
